requirements.txt
python-docx==1.2.0
openpyxl==3.1.5
pandas==2.3.1
ezodf==0.3.2